+++ src/App.js
@@ -1,6 +1,6 @@
 import showMain from './ShowMain.js';
 import MenuButtonController from './MenuButtonController.js';
-import CrewEvent from './CrewEvent.js';
+import CrewEvent from './Crew/CrewEvent.js';
 
 export default class App {
   constructor() {
